requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Werkzeug
email_validator
pytesseract
Pillow
Flask-CSRFProtect
